TicketMonster/src/test/java/org/amdocs/elearning/user/service/UserServiceTest.java
package org.amdocs.elearning.user.service;

import java.time.LocalDate;
import java.util.Optional;

import org.amdocs.elearning.user.service.user.User;
import org.amdocs.elearning.user.service.user.UserDetails;
import org.amdocs.elearning.user.service.user.UserService;
import org.amdocs.elearning.user.service.user.UserType;
import org.junit.Assert;
import org.junit.Test;

public class UserServiceTest {

	final UserService userService = new UserService();

	@Test
	public void getUser_Match() {
		final Optional<User> user = userService.getUserById("0");

		Assert.assertEquals(true, user.isPresent());
	}

	@Test
	public void getUser_NoMatch() {
		final Optional<User> user = userService.getUserById("not a real id");

		Assert.assertEquals(false, user.isPresent());
	}

	@Test
	public void createUser() {
		final UserDetails newUser = new UserDetails("new first", "new last", "M", UserType.PATRON, LocalDate.now());
		final User createdUser = userService.createUser(newUser);
		final Optional<User> retrievedUser = userService.getUserById(createdUser.getId());

		Assert.assertTrue(retrievedUser.isPresent());
		Assert.assertEquals(createdUser, retrievedUser.get());
	}

	@Test
	public void updateUser() {
		final UserDetails updatedUser = new UserDetails("new first", "new last", "M", UserType.PATRON, LocalDate.now());
		final User modifiedUser = userService.updateUser("0", updatedUser);

		final Optional<User> retrievedUser = userService.getUserById("0");
		Assert.assertTrue(retrievedUser.isPresent());
		Assert.assertEquals(modifiedUser, retrievedUser.get());
	}

	@Test
	public void deleteUser() {
		userService.deleteUser("0");
		final Optional<User> retrievedUser = userService.getUserById("0");
		Assert.assertFalse(retrievedUser.isPresent());
	}

}
